make mod.ts simpler
